Fix clippy warnings

+++ src/interface.rs
@@ -246,7 +246,7 @@ impl<'a> Interface<'a> {
                 screen,
                 cursor::MoveTo(
                     1,
-                    (command_line_index as i16 + result_top_index as i16) as u16
+                    (command_line_index + result_top_index as i16) as u16
                 ),
                 SetBackgroundColor(bg),
                 SetForegroundColor(fg),
@@ -310,7 +310,7 @@ impl<'a> Interface<'a> {
                     screen,
                     cursor::MoveTo(
                         width - 9,
-                        (command_line_index as i16 + self.result_top_index() as i16) as u16
+                        (command_line_index + self.result_top_index() as i16) as u16
                     ),
                     SetForegroundColor(timing_color),
                     Print(format!("{:>9}", duration)),
